todo-app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.22
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.25.2
pydantic-settings==2.1.0
openai==1.3.5
pytest==7.4.3
pytest-asyncio==0.21.1
pydantic[email]==2.5.0
asyncpg
python-jose==3.3.0
passlib[bcrypt]

bcrypt==4.0.1
